treewide: refactor `--build-host` to use remote build semantics
Fixes https://github.com/nix-community/nh/issues/428

This is a large architectural change to NH, which lead to me extracting
the remote build logic to its own file so that we may implement it for
Darwin and Home-Manager as well. The `--builders` flag was dropped from
`nh::commands`, and it was replaced with the new and shiny logic that 
hopefully avoids previous pitfalls.

The new `nh::remote` module handles remote builds, including:

- Parsing remote host specifications.
- Copying derivations to remote hosts using `nix-copy-closure`.
- Building derivations on remote hosts via `nix build`.
- Copying results back to localhost or directly to a target host.

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: I236eb1e35dd645f2169462d207bc82e76a6a6964