remote: add SSH reachability checks; enforce local symlink creation
Tiny improvement to how remote connections are made. We now check BEFORE
the connection is made, so that we can avoid all that expensive eval if
it's not reachable. This is not infallible, but it is better. To fix
some target-host quirks, we also have to deal with local symlinks so we
enforce it locally either way.

Signed-off-by: NotAShelf <raf@notashelf.dev>
Change-Id: I65fd7258828459ea82fe6739383567556a6a6964